:recycle: :: 통신적용
